Completed decentralizing tabs, some fixes to the css
